.github/skills/presentation-research/requirements.txt
python-pptx>=0.6.21
requests>=2.0
beautifulsoup4>=4.0
